Moving locale from JavaContainerTest to SshdContainerTest.
